renaming backend to src, adding a docker-compose file along with a Dockerfile. Fixing up imports too
