--- requirements.txt
@@ -1,57 +1,0 @@
-aenum==3.1.15
-aiofiles==24.1.0
-aiohappyeyeballs==2.4.0
-aiohttp==3.10.5
-aiosignal==1.3.1
-annotated-types==0.7.0
-anyio==4.4.0
-attrs==24.2.0
-certifi==2024.8.30
-click==8.1.7
-dataclasses-json==0.6.7
-deepgram-sdk==3.5.1
-deprecation==2.1.0
-distro==1.9.0
-dnspython==2.6.1
-email_validator==2.2.0
-fastapi==0.114.0
-fastapi-cli==0.0.5
-frozenlist==1.4.1
-h11==0.14.0
-httpcore==1.0.5
-httptools==0.6.1
-httpx==0.27.2
-idna==3.8
-Jinja2==3.1.4
-jiter==0.5.0
-loguru==0.7.2
-markdown-it-py==3.0.0
-MarkupSafe==2.1.5
-marshmallow==3.22.0
-mdurl==0.1.2
-multidict==6.1.0
-mypy-extensions==1.0.0
-openai==1.44.1
-packaging==24.1
-pydantic==2.9.1
-pydantic_core==2.23.3
-pydub==0.25.1
-Pygments==2.18.0
-python-dotenv==1.0.1
-python-multipart==0.0.9
-PyYAML==6.0.2
-rich==13.8.1
-setuptools==72.1.0
-shellingham==1.5.4
-sniffio==1.3.1
-starlette==0.38.5
-tqdm==4.66.5
-typer==0.12.5
-typing-inspect==0.9.0
-typing_extensions==4.12.2
-uvicorn==0.30.6
-uvloop==0.20.0
-watchfiles==0.24.0
-websockets==13.0.1
-wheel==0.43.0
-yarl==1.11.1
